Move misc.py to validation.py
