+++ airnow/__init__.py
@@ -7,4 +7,4 @@ from airnow.conditions import *  # noqa: F401, F403
 from airnow.forecast import *  # noqa: F401, F403
 from airnow.historical import *  # noqa: F401, F403
 
-import airnow.misc
+import airnow.validation
